pyproject.toml
[project]
name = "py-script-wrapper"
dynamic = ["version"]
dependencies = [
  "simplejson>=3.16.0",
  "PyYAML>=5.3.1"
]
requires-python = ">=3.8"
authors = [
  { name="Majorant" },
]
maintainers = [
  {name = "Majorant", email="Majorant@users.noreply.github.com"}
]
description = "A simple wrapper for the py-scripts"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License Version 2.0",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Majorant/py-script-wrapper"
Issues = "https://github.com/Majorant/py-script-wrapper/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "swrapper.__version__"}
